move out multiple versions logging tests to new file
